Fix "fixedPoint" formatting with specific precision (T848392) (#11414)
