+++ testing/tests/DevExpress.localization/localization.globalize.tests.js
@@ -47,7 +47,33 @@ define(function(require, exports, module) {
 
     var NBSP = String.fromCharCode(160);
 
-    require('./localization.base.tests.js');
+    var sharedTests = require('./sharedParts/localization.shared.js');
+
+    var NEGATIVE_NUMBERS = [-4.645, -35.855];
+    var ROUNDING_CORRECTION = {
+        '-4.64': '-4.65',
+        '-35.85': '-35.86'
+    };
+    QUnit.module('Globalize common', {
+        before: function() {
+            numberLocalization.inject({
+                format: function(value, format) {
+                    // NOTE: Globalizejs implementation of negative number rounding differs from Intl.
+                    // https://github.com/globalizejs/globalize/issues/884
+                    // If the fractional portion is exactly 0.5 and the argument is negative,
+                    // the argument is rounded to the next integer in the positive direction
+                    var result = this.callBase.apply(this, arguments);
+                    if(NEGATIVE_NUMBERS.indexOf(value) !== -1 && format.type === 'fixedPoint' && format.precision === 2 && !!ROUNDING_CORRECTION[result]) {
+                        result = ROUNDING_CORRECTION[result];
+                    }
+                    return result;
+                }
+            });
+        }
+    }, function() {
+
+        sharedTests();
+    });
 
     QUnit.module('Localization date (ru)', {
         beforeEach: function() {
